pyproject.toml
[project]
name = "sre-bootcamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
]
